Move to avoid confusion with public endpoints
