+++ source/Messaging.Api/MasterDataReceivers/MasterDataResponseMetadata.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Messaging.Api.IncomingMessages;
+namespace Messaging.Api.MasterDataReceivers;
 
 public class MasterDataResponseMetadata
 {
